implement pubsub for shutdown-function
