--- necogcp-app/app.yaml
@@ -1,9 +1,0 @@
-runtime: go116
-
-basic_scaling:
-  max_instances: 1
-  idle_timeout: 1m
-
-handlers:
-  - url: /.*
-    script: auto
